add SelectStatementAssert
