+++ shardingsphere-sql-parser/shardingsphere-sql-parser-test/src/test/java/org/apache/shardingsphere/sql/parser/integrate/asserts/statement/impl/SelectStatementAssert.java
@@ -1,0 +1,108 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.shardingsphere.sql.parser.integrate.asserts.statement.impl;
+
+import com.google.common.base.Optional;
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+import org.apache.shardingsphere.sql.parser.integrate.asserts.SQLCaseAssertContext;
+import org.apache.shardingsphere.sql.parser.integrate.asserts.segment.SQLSegmentAssert;
+import org.apache.shardingsphere.sql.parser.integrate.asserts.segment.groupby.GroupByAssert;
+import org.apache.shardingsphere.sql.parser.integrate.asserts.segment.orderby.OrderByAssert;
+import org.apache.shardingsphere.sql.parser.integrate.asserts.segment.pagination.PaginationAssert;
+import org.apache.shardingsphere.sql.parser.integrate.asserts.segment.predicate.WhereAssert;
+import org.apache.shardingsphere.sql.parser.integrate.asserts.segment.projection.ProjectionAssert;
+import org.apache.shardingsphere.sql.parser.integrate.asserts.segment.table.TableAssert;
+import org.apache.shardingsphere.sql.parser.integrate.jaxb.root.ParserResult;
+import org.apache.shardingsphere.sql.parser.sql.segment.dml.pagination.limit.LimitSegment;
+import org.apache.shardingsphere.sql.parser.sql.statement.dml.SelectStatement;
+
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
+
+/**
+ * Select statement assert.
+ *
+ * @author zhangliang
+ */
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public final class SelectStatementAssert {
+    
+    /**
+     * Assert select statement is correct with expected parser result.
+     * 
+     * @param assertContext Assert context
+     * @param actual actual select statement
+     * @param expected expected parser result
+     */
+    public static void assertSelectStatement(final SQLCaseAssertContext assertContext, final SelectStatement actual, final ParserResult expected) {
+        assertProjection(assertContext, actual, expected);
+        assertTable(assertContext, actual, expected);
+        assertWhere(assertContext, actual, expected);
+        assertGroupBy(assertContext, actual, expected);
+        assertOrderBy(assertContext, actual, expected);
+        assertLimit(assertContext, actual, expected);
+    }
+    
+    private static void assertProjection(final SQLCaseAssertContext assertContext, final SelectStatement actual, final ParserResult expected) {
+        ProjectionAssert.assertIs(assertContext, actual.getProjections(), expected.getProjections());
+    }
+    
+    private static void assertTable(final SQLCaseAssertContext assertContext, final SelectStatement actual, final ParserResult expected) {
+        TableAssert.assertIs(assertContext, actual.getTables(), expected.getTables());
+    }
+    
+    private static void assertWhere(final SQLCaseAssertContext assertContext, final SelectStatement actual, final ParserResult expected) {
+        if (null != expected.getWhere()) {
+            assertTrue(assertContext.getText("Actual where segment should exist."), actual.getWhere().isPresent());
+            WhereAssert.assertIs(assertContext, actual.getWhere().get(), expected.getWhere());
+        } else {
+            assertFalse(assertContext.getText("Actual where segment should not exist."), actual.getWhere().isPresent());
+        }
+    }
+    
+    private static void assertGroupBy(final SQLCaseAssertContext assertContext, final SelectStatement actual, final ParserResult expected) {
+        if (null != expected.getGroupBy()) {
+            assertTrue(assertContext.getText("Actual group by segment should exist."), actual.getGroupBy().isPresent());
+            GroupByAssert.assertIs(assertContext, actual.getGroupBy().get(), expected.getGroupBy());
+        } else {
+            assertFalse(assertContext.getText("Actual group by segment should not exist."), actual.getGroupBy().isPresent());
+        }
+    }
+    
+    private static void assertOrderBy(final SQLCaseAssertContext assertContext, final SelectStatement actual, final ParserResult expected) {
+        if (null != expected.getOrderBy()) {
+            assertTrue(assertContext.getText("Actual order by segment should exist."), actual.getOrderBy().isPresent());
+            OrderByAssert.assertIs(assertContext, actual.getOrderBy().get(), expected.getOrderBy());
+        } else {
+            assertFalse(assertContext.getText("Actual order by segment should not exist."), actual.getOrderBy().isPresent());
+        }
+    }
+    
+    private static void assertLimit(final SQLCaseAssertContext assertContext, final SelectStatement actual, final ParserResult expected) {
+        Optional<LimitSegment> limitSegment = actual.findSQLSegment(LimitSegment.class);
+        if (null != expected.getLimit()) {
+            assertTrue(assertContext.getText("Actual limit segment should exist."), limitSegment.isPresent());
+            PaginationAssert.assertOffset(assertContext, limitSegment.get().getOffset().orNull(), expected.getLimit().getOffset());
+            PaginationAssert.assertRowCount(assertContext, limitSegment.get().getRowCount().orNull(), expected.getLimit().getRowCount());
+            SQLSegmentAssert.assertIs(assertContext, limitSegment.get(), expected.getLimit());
+        } else {
+            assertFalse(assertContext.getText("Actual limit segment should not exist."), limitSegment.isPresent());
+        }
+    }
+}
